Split `MockitoExtension` into traits
