+++ src/main/scala/io/github/effiban/scala2javaext/mockito/providers/MockitoProviders.scala
@@ -1,0 +1,8 @@
+package io.github.effiban.scala2javaext.mockito.providers
+
+import io.github.effiban.scala2java.spi.providers.{AdditionalImportersProvider, ExtendedProviders}
+
+trait MockitoProviders extends ExtendedProviders {
+
+  override def additionalImportersProvider(): AdditionalImportersProvider = MockitoAdditionalImportersProvider
+}
