+++ src/test/scala/io/github/effiban/scala2javaext/mockito/predicate/MockitoPredicatesTest.scala
@@ -1,0 +1,17 @@
+package io.github.effiban.scala2javaext.mockito.predicate
+
+import io.github.effiban.scala2javaext.mockito.testsuites.UnitTestSuite
+
+class MockitoPredicatesTest extends UnitTestSuite {
+
+  private val mockitoPredicates = new MockitoPredicates {}
+  import mockitoPredicates._
+
+  test("importerExcludedPredicate() should return MockitoImporterExcludedPredicate") {
+    importerExcludedPredicate() shouldBe MockitoImporterExcludedPredicate
+  }
+
+  test("templateInitExcludedPredicate() should return MockitoTemplateInitExcludedPredicate") {
+    templateInitExcludedPredicate() shouldBe MockitoTemplateInitExcludedPredicate
+  }
+}
