+++ src/test/scala/io/github/effiban/scala2javaext/mockito/transformer/MockitoTransformersTest.scala
@@ -1,0 +1,37 @@
+package io.github.effiban.scala2javaext.mockito.transformer
+
+import io.github.effiban.scala2javaext.mockito.testsuites.UnitTestSuite
+
+class MockitoTransformersTest extends UnitTestSuite {
+
+  private val mockitoTransformers = new MockitoTransformers {}
+  import mockitoTransformers._
+
+  test("importerTransformer() should return MockitoImporterTransformer") {
+    importerTransformer() shouldBe MockitoImporterTransformer
+  }
+
+  test("classTransformer() should return MockitoClassTransformer") {
+    classTransformer() shouldBe MockitoClassTransformer
+  }
+
+  test("defnValTransformer() should return MockitoDefnValTransformer") {
+    defnValTransformer() shouldBe MockitoDefnValTransformer
+  }
+
+  test("defnValToDeclVarTransformer() should return MockitoDefnValToDeclVarTransformer") {
+    defnValToDeclVarTransformer() shouldBe MockitoDefnValToDeclVarTransformer
+  }
+
+  test("termApplyTypeToTermApplyTransformer() should return MockitoTermApplyTypeToTermApplyTransformer") {
+    termApplyTypeToTermApplyTransformer() shouldBe MockitoTermApplyTypeToTermApplyTransformer
+  }
+
+  test("termApplyTransformer() should return MockitoTermApplyTransformer") {
+    termApplyTransformer() shouldBe MockitoTermApplyTransformer
+  }
+
+  test("termSelectTransformer() should return MockitoTermSelectTransformer") {
+    termSelectTransformer() shouldBe MockitoTermSelectTransformer
+  }
+}
